refactor java code gen into a separate object called CodeGenJava under package codegen
